+++ src/main/java/com/kakaotech/team14backend/oauth2/application/CreateKakaoUser.java
@@ -1,0 +1,50 @@
+package com.kakaotech.team14backend.oauth2.application;
+
+
+import com.kakaotech.team14backend.auth.PrincipalDetails;
+import com.kakaotech.team14backend.inner.member.model.Member;
+import com.kakaotech.team14backend.inner.member.model.Role;
+import com.kakaotech.team14backend.inner.member.model.Status;
+import com.kakaotech.team14backend.inner.member.repository.MemberRepository;
+import com.kakaotech.team14backend.oauth2.dto.KakaoProfileDTO;
+import com.kakaotech.team14backend.outer.member.service.MemberService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
+
+@Component
+@RequiredArgsConstructor
+public class CreateKakaoUser {
+  private final MemberService memberService;
+  private final MemberRepository memberRepository;
+  @Transactional
+  public void execute(KakaoProfileDTO kakaoProfileDTO) {
+    String kakaoId = kakaoProfileDTO.getId();
+    String userName = kakaoProfileDTO.getProperties().getNickname();
+    String profileImage = kakaoProfileDTO.getProperties().getProfileImage();
+    Member memberEntity = memberRepository.findByKakaoId(kakaoId);
+
+    if (memberEntity == null) {
+      memberEntity = memberService.createMember(
+          userName,
+          kakaoId,
+          "none",
+          profileImage,
+          Role.ROLE_BEGINNER,
+          0L,
+          Status.STATUS_ACTIVE);
+      memberRepository.save(memberEntity);
+    }
+    if (memberEntity.getUserStatus().equals(Status.STATUS_INACTIVE)) {
+      memberService.makeUserActive(memberEntity.getMemberId());
+    }
+
+    PrincipalDetails principalDetails = new PrincipalDetails(memberEntity);
+
+    Authentication authentication = new UsernamePasswordAuthenticationToken(principalDetails, null, principalDetails.getAuthorities());
+    SecurityContextHolder.getContext().setAuthentication(authentication);
+  }
+}
