+++ src/main/java/com/kakaotech/team14backend/oauth2/dto/GetInstagramCodeDTO.java
@@ -1,0 +1,8 @@
+package com.kakaotech.team14backend.oauth2.dto;
+
+import lombok.Data;
+
+@Data
+public class GetInstagramCodeDTO {
+  String code;
+}
