+++ src/main/java/com/kakaotech/team14backend/oauth2/dto/GetKakaoCodeDTO.java
@@ -1,0 +1,8 @@
+package com.kakaotech.team14backend.oauth2.dto;
+
+import lombok.Data;
+
+@Data
+public class GetKakaoCodeDTO {
+  private String code;
+}
