--- src/main/java/com/kakaotech/team14backend/outer/login/dto/GetInstagramCode.java
@@ -1,8 +1,0 @@
-package com.kakaotech.team14backend.outer.login.dto;
-
-import lombok.Data;
-
-@Data
-public class GetInstagramCode {
-  String code;
-}
