--- src/main/java/com/kakaotech/team14backend/outer/login/dto/GetKakaoCode.java
@@ -1,8 +1,0 @@
-package com.kakaotech.team14backend.outer.login.dto;
-
-import lombok.Data;
-
-@Data
-public class GetKakaoCode {
-  private String code;
-}
